Add batch capabilities
