+++ modules/text2vec-ollama/ent/class_settings.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package vectorizer
+package ent
 
 import (
 	"github.com/pkg/errors"
