--- modules/text2vec-ollama/vectorizer/objects.go
@@ -1,74 +1,0 @@
-//                           _       _
-// __      _____  __ ___   ___  __ _| |_ ___
-// \ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
-//  \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
-//   \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
-//
-//  Copyright © 2016 - 2024 Weaviate B.V. All rights reserved.
-//
-//  CONTACT: hello@weaviate.io
-//
-
-package vectorizer
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/weaviate/weaviate/entities/models"
-	"github.com/weaviate/weaviate/entities/moduletools"
-	"github.com/weaviate/weaviate/usecases/modulecomponents"
-	objectsvectorizer "github.com/weaviate/weaviate/usecases/modulecomponents/vectorizer"
-)
-
-type Vectorizer struct {
-	client           Client
-	objectVectorizer *objectsvectorizer.ObjectVectorizer
-}
-
-func New(client Client) *Vectorizer {
-	return &Vectorizer{
-		client:           client,
-		objectVectorizer: objectsvectorizer.New(),
-	}
-}
-
-type Client interface {
-	Vectorize(ctx context.Context, input []string,
-		cfg moduletools.ClassConfig,
-	) (*modulecomponents.VectorizationResult, *modulecomponents.RateLimits, int, error)
-	VectorizeQuery(ctx context.Context, input []string,
-		cfg moduletools.ClassConfig,
-	) (*modulecomponents.VectorizationResult, error)
-}
-
-// IndexCheck returns whether a property of a class should be indexed
-type ClassSettings interface {
-	PropertyIndexed(property string) bool
-	VectorizePropertyName(propertyName string) bool
-	VectorizeClassName() bool
-	ApiEndpoint() string
-	ProjectID() string
-	ModelID() string
-	TitleProperty() string
-}
-
-func (v *Vectorizer) Object(ctx context.Context, object *models.Object, cfg moduletools.ClassConfig,
-) ([]float32, models.AdditionalProperties, error) {
-	vec, err := v.object(ctx, object, cfg)
-	return vec, nil, err
-}
-
-func (v *Vectorizer) object(ctx context.Context, object *models.Object, cfg moduletools.ClassConfig,
-) ([]float32, error) {
-	icheck := NewClassSettings(cfg)
-	text := v.objectVectorizer.Texts(ctx, object, icheck)
-	res, _, _, err := v.client.Vectorize(ctx, []string{text}, cfg)
-	if err != nil {
-		return nil, err
-	}
-	if len(res.Vector) != 1 {
-		return nil, fmt.Errorf("got %d embeddings back but should get only 1", len(res.Vector))
-	}
-	return res.Vector[0], nil
-}
