--- modules/text2vec-ollama/vectorizer/texts_test.go
@@ -1,47 +1,0 @@
-//                           _       _
-// __      _____  __ ___   ___  __ _| |_ ___
-// \ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
-//  \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
-//   \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
-//
-//  Copyright © 2016 - 2024 Weaviate B.V. All rights reserved.
-//
-//  CONTACT: hello@weaviate.io
-//
-
-package vectorizer
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-// as used in the nearText searcher
-func TestVectorizingTexts(t *testing.T) {
-	type testCase struct {
-		name  string
-		input []string
-	}
-
-	tests := []testCase{
-		{
-			name:  "single word",
-			input: []string{"hello"},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			client := &fakeClient{}
-
-			v := New(client)
-			vec, err := v.Texts(context.Background(), test.input, &fakeClassConfig{})
-
-			require.Nil(t, err)
-			assert.Equal(t, []float32{0, 1, 2, 3}, vec)
-		})
-	}
-}
